added tanstack query

+++ finance_manager_frontend/categories/CategoriesAPI.ts
@@ -2,22 +2,41 @@ import { CategoryEntity } from "./CategoryEntity";
 import axios from "axios";
 import * as SecureStore from "expo-secure-store";
 
-
 export class CategoriesAPI {
-  static baseUrl = "http://10.59.169.159:3000/categories";
+  static baseUrl = "http://192.168.0.19:3000/categories";
+  
+  
+    static async getCategories() {
+      try {
+        const token = await SecureStore.getItemAsync("jwt");
+        if (!token) {
+          throw new Error("No token found");
+        }
+  
+        console.log("Token:", token);
+        console.log("Calling GET endpoint:", `${this.baseUrl}/user`);
+  
+        const response = await axios.get(`${this.baseUrl}/user`, {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+  
+        console.log("Response data:", response.data);
+        return response.data;
+      } catch (error) {
+        console.error("Error fetching categories:", error);
+  
+        // If the error is an Axios error, log the response
+        if (axios.isAxiosError(error)) {
+          console.error("Axios error response:", error.response?.data);
+        }
+  
+        // Rethrow the error to let useQuery handle it
+        throw error;
+      }
+    }
 
-  static async getCategories() {
-    const token = await SecureStore.getItemAsync("jwt");
-    console.log("calling " + CategoriesAPI.baseUrl + "/user");
-
-    const response = await axios.get<CategoryEntity[]>(this.baseUrl + "/user", {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    return response.data;
-  }
 
   static async createCategory(category: CategoryEntity) {
     console.log("calling " + CategoriesAPI.baseUrl);
@@ -57,6 +76,11 @@ export class CategoriesAPI {
 
   static async deleteCategory(id: number) {
     const token = await SecureStore.getItemAsync("jwt");
+    console.log("token", token);
+    console.log("id", id);
+    if (!token) {
+      throw new Error("No token found");
+    }
     console.log("calling " + CategoriesAPI.baseUrl + "/" + id);
 
     const response = await axios.delete(this.baseUrl + "/" + id, {
@@ -67,5 +91,4 @@ export class CategoriesAPI {
 
     return response.data;
   }
-
 }
